chore: split engine_shard file from engine_shard_set
No functional changes besides that.
