+++ src/core/intent_lock.h
@@ -3,6 +3,8 @@
 //
 #include <assert.h>
 
+#include <ostream>
+
 #pragma once
 
 namespace dfly {
